rename helpers.py to baroque_project.py
